+++ client-app/shared/common/types/extension-registry.ts
